[fix] Do not create .htaccess with Nginx

+++ lib/Persistence/AbstractPersistence.php
@@ -82,21 +82,23 @@ abstract class AbstractPersistence
      */
     protected static function _initialize()
     {
-        // Create storage directory if it does not exist.
-        if (!is_dir(self::$_path)) {
-            if (!@mkdir(self::$_path, 0700)) {
-                throw new Exception('unable to create directory ' . self::$_path, 10);
+        if (property_exists($data->meta, 'webserver') && $data->meta->webserver && $this->_conf->getKey('webserver') == "Apache") {
+            // Create storage directory if it does not exist.
+            if (!is_dir(self::$_path)) {
+                if (!@mkdir(self::$_path, 0700)) {
+                    throw new Exception('unable to create directory ' . self::$_path, 10);
+                }
             }
-        }
-        $file = self::$_path . DIRECTORY_SEPARATOR . '.htaccess';
-        if (!is_file($file)) {
-            $writtenBytes = @file_put_contents(
-                $file,
-                'Require all denied' . PHP_EOL,
-                LOCK_EX
-            );
-            if ($writtenBytes === false || $writtenBytes < 19) {
-                throw new Exception('unable to write to file ' . $file, 11);
+            $file = self::$_path . DIRECTORY_SEPARATOR . '.htaccess';
+            if (!is_file($file)) {
+                $writtenBytes = @file_put_contents(
+                    $file,
+                    'Require all denied' . PHP_EOL,
+                    LOCK_EX
+                );
+                if ($writtenBytes === false || $writtenBytes < 19) {
+                    throw new Exception('unable to write to file ' . $file, 11);
+                }
             }
         }
     }
